% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{add_gpos}
\alias{add_gpos}
\title{Add genomic position}
\usage{
add_gpos(tib, ...)
}
\description{
\code{add_gpos} adds the genomic position to a tibble of genomic stats.
}
\details{
The imported data typically comes with one column specifieing the linkage group (LG)
and one specifieng the position of that LG.
This function translates the postion into the genomic postion where
all LGs have been concatinated (eg. LG02, POS:1 => length(LG01)+1)
}
