% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF7.R
\name{add_spec_drawing}
\alias{add_spec_drawing}
\title{Add fish annotation}
\usage{
add_spec_drawing(spec, pos)
}
\description{
\code{add_spec_drawing} adds a fish annotation to a plot
}
\seealso{
Other Suppl Figure 7: 
\code{\link{adm_plot}()},
\code{\link{data_amdx}()},
\code{\link{read_sc}()}
}
\concept{Suppl Figure 7}
