% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX9.R
\name{adm_plot}
\alias{adm_plot}
\title{Plot adimxture data}
\usage{
adm_plot(gid_in)
}
\description{
\code{adm_plot} plots the admixture data
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{collapse_peaks}()},
\code{\link{data_amdx}()},
\code{\link{getPofZ}()},
\code{\link{get_gxp_both_models}()},
\code{\link{label_spacer}()},
\code{\link{plot_loc}()},
\code{\link{reformat_run_name}()},
\code{\link{theme_hyb}()}
}
\concept{Figure SX}
