% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX3.R
\name{anno_pair_flag}
\alias{anno_pair_flag}
\title{create hamlet pair annotations}
\usage{
anno_pair_flag(
  loc,
  left,
  right,
  circle_color = NA,
  circle_fill_left = "white",
  circle_fill_right = "lightgray",
  circle_lwd = 0.5,
  plot_names = FALSE,
  plot_name_size = 3,
  font_family = "sans",
  ...
)
}
\description{
\code{plot_pair_run} creates the hamlet pair annotations.
}
\seealso{
Other Suppl. Figure SX: 
\code{\link{StatTreeData}},
\code{\link{fish_plot2}()},
\code{\link{fish_plot}()},
\code{\link{fst_bar_row_run}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{plot_fishes_location}()},
\code{\link{plot_pair_run}()},
\code{\link{refactor_run}()},
\code{\link{rescale_fst}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()},
\code{\link{summarise_model}()}
}
\concept{Suppl. Figure SX}
