% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tables.R
\name{as_multirow}
\alias{as_multirow}
\title{Wrap content in multirow}
\usage{
as_multirow(x, n)
}
\arguments{
\item{x}{string, cell content}

\item{n}{integer, number of rows to combine}
}
\description{
\code{as_multirow} formats multirow content for latex tables.
}
\seealso{
Other Table functions: 
\code{\link{get_genes}()}
}
\concept{Table functions}
