% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{clr}
\alias{clr}
\title{Hamlet colors (standard)}
\format{
An object of class \code{character} of length 11.
}
\usage{
clr
}
\description{
\code{clr} defines the standard colors for hamlets species within plots.
}
\details{
This is used in most color scales thoughout the paper
where individual species are indicated by color.
}
\keyword{datasets}
