% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{clr2}
\alias{clr2}
\title{Hamlet colors (alternative)}
\format{An object of class \code{character} of length 11.}
\usage{
clr2
}
\description{
\code{clr2} defines the alternative colors for hamlets species within plots.
}
\details{
This is used in color scales where cloring H. unicolor
white does not work.
}
\keyword{datasets}
