% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{clr_below}
\alias{clr_below}
\title{Default background bar color}
\format{
An object of class \code{character} of length 1.
}
\usage{
clr_below
}
\description{
\code{clr_below} is the default background bar color.
}
\keyword{datasets}
