% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{clr_genes}
\alias{clr_genes}
\title{Default gene color}
\format{
An object of class \code{character} of length 1.
}
\usage{
clr_genes
}
\description{
\code{clr_genes} is the default color of genes zoom plot.w
}
\keyword{datasets}
