% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{clr_loc}
\alias{clr_loc}
\title{Location colors}
\format{An object of class \code{character} of length 4.}
\usage{
clr_loc
}
\description{
\code{clr_loc} defines the colors for sample locations within plots.
}
\details{
This is used in color scales thoughout the paper
where the locations are indicated by color.
}
\keyword{datasets}
