% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{collapse_peaks}
\alias{collapse_peaks}
\title{Collapse all Fst peaks of a given run}
\usage{
collapse_peaks(x)
}
\arguments{
\item{x}{string vector with outlier peak ids}
}
\description{
\code{collapse_peaks} collapses all Fst peaks of a given run
}
