% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_post_review.R
\name{conditional_highlight}
\alias{conditional_highlight}
\title{Conditional Tree Highlighting}
\usage{
conditional_highlight(tree, highl = TRUE, higl_node, support_guide = FALSE)
}
\arguments{
\item{tree}{tree object}

\item{highl}{logical, should there be highlighting}

\item{higl_node}{node, where to start highlighting}

\item{support_guide}{logical, show legend for node support}
}
\description{
Conditional Tree Highlighting
}
