% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_post_review.R
\name{conditional_rotate}
\alias{conditional_rotate}
\title{Conditional Tree Rotation}
\usage{
conditional_rotate(tree, rotate = TRUE, root_node)
}
\arguments{
\item{tree}{tree object}

\item{rotate}{logical, should the tree be rotated}

\item{root_node}{node, where to rotate the tree}
}
\description{
Conditional Tree Rotation
}
