% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{darken}
\alias{darken}
\title{Darken a given color}
\usage{
darken(color, factor = 0.4)
}
\description{
\code{darken} makes a darker version of a given color.
}
\seealso{
Other General functions: 
\code{\link{export_2_latex}()},
\code{\link{get_dxy}()},
\code{\link{get_fst_fixed}()},
\code{\link{get_fst_summary_data}()},
\code{\link{get_fst}()},
\code{\link{get_gxp}()},
\code{\link{get_pi}()},
\code{\link{join_list}()},
\code{\link{lighten}()},
\code{\link{process_input}()},
\code{\link{refactor}()}
}
\concept{General functions}
