% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF7.R
\name{data_amdx}
\alias{data_amdx}
\title{Get adimxture data}
\usage{
data_amdx(gid, k, admx_path)
}
\description{
\code{data_amdx} imports the admixture data
}
\seealso{
Other Suppl Figure 7: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{read_sc}()}
}
\concept{Suppl Figure 7}
