% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cross_topo_f3.R
\name{distances_tree}
\alias{distances_tree}
\title{Get distance matrix of a topology}
\usage{
distances_tree(tree)
}
\description{
\code{distances_tree} unroots a topology and creates a distance matrix from it
}
\seealso{
Other Figure 3: \code{\link{ax_scl}},
  \code{\link{cross_spec}}, \code{\link{custom_annoplot}},
  \code{\link{dist_tibble}}, \code{\link{get_clr}},
  \code{\link{get_dist}}, \code{\link{get_gxp_long}},
  \code{\link{get_isolated_topos}},
  \code{\link{get_neighbour_topos}},
  \code{\link{is_isolated}}, \code{\link{min_dist}},
  \code{\link{no_title}}, \code{\link{plot_curtain}},
  \code{\link{plot_fish_zoom}}, \code{\link{plot_leg}},
  \code{\link{plot_panel_anno}},
  \code{\link{plot_panel_delta_dxy}},
  \code{\link{plot_panel_dxy}},
  \code{\link{plot_panel_fst}},
  \code{\link{plot_panel_gxp}},
  \code{\link{plot_panel_twisst}}, \code{\link{prep_data}},
  \code{\link{theme_panels}}
}
\concept{Figure 3}
