% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.onAttach}
\alias{.onAttach}
\title{Package Dependencies}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{dummy}

\item{pkgname}{dummy}
}
\description{
Added so roxygen2 adds specific packages to NAMESPACE
}
