% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{export_2_latex}
\alias{export_2_latex}
\title{Export latex tables}
\usage{
export_2_latex(table, name)
}
\description{
\code{export_2_latex} exports a formated latex table.
}
\seealso{
Other General functions: \code{\link{darken}},
  \code{\link{get_dxy}}, \code{\link{get_fst_fixed}},
  \code{\link{get_fst}}, \code{\link{get_gxp}},
  \code{\link{get_pi}}, \code{\link{join_list}},
  \code{\link{lighten}}, \code{\link{process_input}},
  \code{\link{refactor}}
}
\concept{General functions}
