% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{f_width_half}
\alias{f_width_half}
\title{Default figure width (single column)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
f_width_half
}
\description{
\code{f_width_half} defines the half width for figures (mm).
}
\details{
3.43 inch = 87mm
}
\keyword{datasets}
