% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF10.R
\name{fish_plot}
\alias{fish_plot}
\title{hamlet pair annotation}
\usage{
fish_plot(spec)
}
\arguments{
\item{spec}{string six letter population name}
}
\description{
\code{fish_plot} provides hamlet pair annotation.
}
