% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX4.R
\name{fish_plot2}
\alias{fish_plot2}
\title{hamlet pair annotation 2}
\usage{
fish_plot2(spec)
}
\description{
\code{fish_plot2} provides hamlet pair annotation with alternative axis specifications.
}
\seealso{
Other Suppl. Figure SX: \code{\link{StatTreeData}},
  \code{\link{anno_pair_flag}}, \code{\link{fish_plot}},
  \code{\link{fst_bar_row_run}},
  \code{\link{geom_hypo_grob2}},
  \code{\link{geom_nodepoint_support}},
  \code{\link{get_tree}},
  \code{\link{plot_fishes_location}},
  \code{\link{plot_pair_run}}, \code{\link{refactor_run}},
  \code{\link{rescale_fst}}, \code{\link{root_manual}},
  \code{\link{setup_tree_data}},
  \code{\link{summarise_model}}
}
\concept{Suppl. Figure SX}
