% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{fll_fun}
\alias{fll_fun}
\title{Fill function for adxiture classes}
\usage{
fll_fun(n)
}
\arguments{
\item{n}{integer, number of clusters}
}
\description{
\code{fll_fun} creates a fill scale for admixture classes.
}
\details{
This is used to keep a consistent color scheme thoughout
various admixture plots of different k.
}
