% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{fll_n}
\alias{fll_n}
\title{Named fill function for adxiture classes}
\usage{
fll_n(n)
}
\arguments{
\item{n}{integer, number of clusters}
}
\description{
\code{fll_n} creates a named fill scale for admixture classes.
}
\details{
This is used to keep a consistent color scheme thoughout
various admixture plots of different k.
The named variant is used to assign the colors consistently.
}
