% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX3.R
\name{fst_bar_row_run}
\alias{fst_bar_row_run}
\title{crate global fst bar table}
\usage{
fst_bar_row_run(fst, run)
}
\description{
\code{fst_bar_row_run} crates a rescaled global fst bar table.
}
\seealso{
Other Suppl. Figure SX: \code{\link{StatTreeData}},
  \code{\link{anno_pair_flag}}, \code{\link{fish_plot2}},
  \code{\link{fish_plot}}, \code{\link{geom_hypo_grob2}},
  \code{\link{geom_nodepoint_support}},
  \code{\link{get_tree}},
  \code{\link{plot_fishes_location}},
  \code{\link{plot_pair_run}}, \code{\link{refactor_run}},
  \code{\link{rescale_fst}}, \code{\link{root_manual}},
  \code{\link{setup_tree_data}},
  \code{\link{summarise_model}}
}
\concept{Suppl. Figure SX}
