% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generic_hamlet_img}
\alias{generic_hamlet_img}
\title{Generic Hamlet image.}
\format{
A ggtree object with a generic hamlet image:
}
\source{
KH drawing
}
\usage{
generic_hamlet_img
}
\description{
A generic hamlet image
}
\keyword{datasets}
