% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF3.R
\name{geom_hypo_grob2}
\alias{geom_hypo_grob2}
\title{The custom grob geom}
\usage{
geom_hypo_grob2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\description{
This geom builds heavily on the answer by baptiste on the
tidiverse github forum:
https://github.com/tidyverse/ggplot2/issues/1399
}
\seealso{
Other Suppl Figure 3: 
\code{\link{collapse_peaks}()},
\code{\link{get_fst_fixed}()},
\code{\link{plot_fishes_location}()},
\code{\link{reformat_run_name}()},
\code{\link{summarise_model}()}
}
\concept{Suppl Figure 3}
