% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF6.R
\name{geom_nodepoint_support}
\alias{geom_nodepoint_support}
\title{backbone for ggplot layer for node support}
\usage{
geom_nodepoint_support(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
\code{geom_nodepoint_support} provides the backbone for a custom ggplot layer for node support.
}
\seealso{
Other Suppl Figure 6: 
\code{\link{StatTreeData}},
\code{\link{get_tree}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()}
}
\concept{Suppl Figure 6}
