% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f2.R
\name{get_cc}
\alias{get_cc}
\title{Import cross coalescence data}
\usage{
get_cc(file, cc_groups, cc_path, mu = 3.7e-08, gen = 1)
}
\arguments{
\item{file}{cross coalescence file}

\item{cc_groups}{cross coalescence group assignment reference table}

\item{cc_path}{cross coalescence results path}

\item{mu}{mutation rate mu}

\item{gen}{generation time in years}
}
\description{
\code{get_msmc} imports cross coalescence data and extracts according run information.
}
\details{
The data is imported and run components are extracted from the a reference table,
and added to the data.
}
\seealso{
Other Figure 2: 
\code{\link{get_msmc}()}
}
\concept{Figure 2}
