% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{get_dxy}
\alias{get_dxy}
\title{Import dxy data}
\usage{
get_dxy(file)
}
\description{
\code{get_dxy} inports dxy data, adds the genomic position and the run.
}
\details{
The run name (the pair wise species comparison) is extracted from the file name,
the data is imported, column names are standardized and genomic position and
run name are added.
}
