% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{get_fst}
\alias{get_fst}
\title{Import Fst data}
\usage{
get_fst(file, kb = "50k")
}
\description{
\code{get_fst} imports Fst data, adds the genomic position and the run.
}
\details{
The run name (the pair wise species comparison) is extracted from the file name,
the data is imported, column names are standardized and genomic position and
run name are added.
}
\seealso{
Other General functions: \code{\link{darken}},
  \code{\link{get_dxy}}, \code{\link{get_gxp}},
  \code{\link{get_pi}}, \code{\link{join_list}},
  \code{\link{lighten}}, \code{\link{process_input}},
  \code{\link{refactor}}
}
\concept{General functions}
