% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{get_fst_summary_data}
\alias{get_fst_summary_data}
\title{Import fst summary data for outlier windows}
\usage{
get_fst_summary_data(gid)
}
\description{
\code{get_fst_summary_data} imports pair wise fst summary data for outlier windows.
}
\seealso{
Other General functions: 
\code{\link{darken}()},
\code{\link{export_2_latex}()},
\code{\link{get_dxy}()},
\code{\link{get_fst}()},
\code{\link{get_gxp}()},
\code{\link{get_pi}()},
\code{\link{join_list}()},
\code{\link{lighten}()},
\code{\link{process_input}()},
\code{\link{refactor}()}
}
\concept{General functions}
