% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{get_gxp}
\alias{get_gxp}
\title{Import genotype x phenotype association}
\usage{
get_gxp(file)
}
\description{
\code{get_gxp} imports genotype x phenotype association data and adds the genomic position.
}
\details{
The phenotype trait name is extracted from the file name and
converted into a label for plotting.
Then, the data is imported and reduced to the columns containg
the genomic poistion and the p values of the wald test.
The column containing the p values is renamed according to the trait.
}
\seealso{
Other General functions: \code{\link{get_dxy}},
  \code{\link{get_fst}}, \code{\link{join_list}},
  \code{\link{process_input}}
}
\concept{General functions}
