% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f6.R
\name{get_gxp_and_geva}
\alias{get_gxp_and_geva}
\title{Combine allele age and gxp data import}
\usage{
get_gxp_and_geva(gid, chrom, start, end, ...)
}
\arguments{
\item{gid}{string, focal outlier id}

\item{chrom}{string, linkage group identifier}

\item{start}{numeric, start position (bp)}

\item{end}{numeric, end position (bp)}

\item{...}{catch all parameter, for usage with purrr::pmap() and unexpected parameters}
}
\description{
Combine allele age and gxp data import
}
