% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX11.R
\name{get_gxp_both_models}
\alias{get_gxp_both_models}
\title{Get gxp data}
\usage{
get_gxp_both_models(file, trait, model_type, path)
}
\description{
\code{get_gxp_both_models} imports gxp data
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{collapse_peaks}()},
\code{\link{data_amdx}()},
\code{\link{getPofZ}()},
\code{\link{plot_loc}()},
\code{\link{reformat_run_name}()}
}
\concept{Figure SX}
