% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{get_gxp_long}
\alias{get_gxp_long}
\title{Import genotype x phenotype association in long format}
\usage{
get_gxp_long(file, kb = 10)
}
\description{
\code{get_gxp_long} imports genotype x phenotype association data in long format.
}
\details{
The phenotype trait name is extracted from the file name.
Then, the data is imported, the genomic poistion and the trait name is added as columns.
}
\seealso{
Other Figure 3: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{no_title}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 3}
