% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f2.R
\name{get_msmc}
\alias{get_msmc}
\title{Import msmc data}
\usage{
get_msmc(file, msmc_path, mu = 3.7e-08, gen = 1)
}
\arguments{
\item{file}{msmc file}

\item{msmc_path}{msmc results path}

\item{mu}{mutation rate mu}

\item{gen}{generation time in years}
}
\description{
\code{get_msmc} imports msc data.
}
\details{
The run name (the pair wise species comparison) is extracted from the file name,
the data is imported and the run name is added.
}
\seealso{
Other Figure 2: 
\code{\link{get_cc}()}
}
\concept{Figure 2}
