% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF15.R
\name{get_n_above}
\alias{get_n_above}
\title{Count number of permutation Fst greather than the original Fst}
\usage{
get_n_above(data)
}
\arguments{
\item{data}{tibble}
}
\description{
Count number of permutation Fst greather than the original Fst
}
