% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_cross_topo.R
\name{get_neighbour_topos}
\alias{get_neighbour_topos}
\title{Get pairwise highlighing topologies}
\usage{
get_neighbour_topos(topo_plots, left_neighbour, right_neighbour)
}
\description{
\code{get_neighbour_topos} finds all topologies where two species are sister species
}
\details{
First the distances of all species within all topologies are created.
This is subset for the topologies where the distance between the two species equals 2 (2 edges).
}
\seealso{
Other Figure 5: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{no_title}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 5}
