% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF15.R
\name{get_percentile}
\alias{get_percentile}
\title{Compute the percentile of original Fst copared to fst distribution}
\usage{
get_percentile(data)
}
\arguments{
\item{data}{tibble}
}
\description{
Compute the percentile of original Fst copared to fst distribution
}
