% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{get_pi}
\alias{get_pi}
\title{Import diversity data}
\usage{
get_pi(file)
}
\description{
\code{get_pi} imports diversity data and adds the genomic position.
}
\details{
The data is imported and genomic position and species name are added.
}
\seealso{
Other General functions: 
\code{\link{darken}()},
\code{\link{export_2_latex}()},
\code{\link{get_dxy}()},
\code{\link{get_fst_summary_data}()},
\code{\link{get_fst}()},
\code{\link{get_gxp}()},
\code{\link{join_list}()},
\code{\link{lighten}()},
\code{\link{process_input}()},
\code{\link{refactor}()}
}
\concept{General functions}
