% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF15.R
\name{get_random_fst}
\alias{get_random_fst}
\title{Import Fst permutation tests results}
\usage{
get_random_fst(file)
}
\arguments{
\item{file}{string, path to fst permutation test results}
}
\description{
Import Fst permutation tests results
}
