% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF6.R
\name{get_tree}
\alias{get_tree}
\title{import phylogenetic tree}
\usage{
get_tree(loc, file, tree_dir, ...)
}
\description{
\code{get_tree} imports a phylogenetic tree.
}
\seealso{
Other Suppl Figure 6: 
\code{\link{StatTreeData}},
\code{\link{geom_nodepoint_support}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()}
}
\concept{Suppl Figure 6}
