% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f4.R
\name{get_twisst_data}
\alias{get_twisst_data}
\title{Import a single topology weighting data set}
\usage{
get_twisst_data(loc, w_in, d_in, smooth = FALSE, span = 0.01)
}
\arguments{
\item{loc}{sample location (bel [Belize]/ hon [Honduras]/ pan [Panama])}

\item{w_in}{weights file (topology weighting results)}

\item{d_in}{data file (window position)}

\item{smooth}{logical - apply smoothing?}

\item{span}{granularity of smoothing}
}
\description{
\code{get_twisst_data} a single topology weighting data set
}
\details{
The topology weighting file containg the weights is read,
normalized and filtered for missing data.
Then the file containg the window positions is read,
filtered according to the weights file and its columns
are renamed.

Optionally, the data is smoothed, then the weights and
the position information is merged, reformated and the
genomic positions are added.
}
\seealso{
Other Figure 4: 
\code{\link{match_twisst_files}()}
}
\concept{Figure 4}
