% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f6.R
\name{grid_piece_x}
\alias{grid_piece_x}
\title{Single grob annotation for specific x range}
\usage{
grid_piece_x(g, x1, x2, ...)
}
\arguments{
\item{g}{grob oject}

\item{x1}{numeric, left border}

\item{x2}{numeric, right border}

\item{...}{catch all parameter, for usage with purrr::pmap() and unexpected parameters}
}
\description{
Single grob annotation for specific x range
}
