% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f6.R
\name{gxp_importer}
\alias{gxp_importer}
\title{Import windowed gxp data for focal trait}
\usage{
gxp_importer(trait, chrom, start, end)
}
\arguments{
\item{trait}{string, gxp focal trait}

\item{chrom}{string, linkage group identifier}

\item{start}{numeric, start position (bp)}

\item{end}{numeric, end position (bp)}
}
\description{
Import windowed gxp data for focal trait
}
