% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f6.R
\name{import_geva_data}
\alias{import_geva_data}
\title{Import Geva data (Allele Age)}
\usage{
import_geva_data(chrom, start, end, ...)
}
\arguments{
\item{chrom}{string, linkage group identifier}

\item{start}{numeric, start position (bp)}

\item{end}{numeric, end position (bp)}

\item{...}{catch all parameter, for usage with purrr::pmap() and unexpected parameters}
}
\description{
Import Geva data (Allele Age)
}
