% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{lighten}
\alias{lighten}
\title{Lighten a given color}
\usage{
lighten(color, factor = 0.4)
}
\description{
\code{lighten} makes a lighter version of a given color.
}
\seealso{
Other General functions: \code{\link{darken}},
  \code{\link{get_dxy}}, \code{\link{get_fst}},
  \code{\link{get_gxp}}, \code{\link{join_list}},
  \code{\link{process_input}}, \code{\link{refactor}}
}
\concept{General functions}
