% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{loc_names}
\alias{loc_names}
\title{Location names}
\format{
An object of class \code{character} of length 4.
}
\usage{
loc_names
}
\description{
\code{loc_names} stores the geographic names of the sample abrreviations.
}
\details{
This is used to translate between the 3-letter sample abrreviations and
the actual geographic names of the sample locations.
}
\keyword{datasets}
