% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{make_faint_clr}
\alias{make_faint_clr}
\title{make a faint version of a color}
\usage{
make_faint_clr(loc)
}
\description{
\code{make_faint_clr} creates a faint version of a given color.
}
\seealso{
Other Figure 1: \code{\link{network_layout}},
  \code{\link{plot_fish}}, \code{\link{plot_network}},
  \code{\link{summarize_dxy}}, \code{\link{summarize_fst}},
  \code{\link{whisker_points}}
}
\concept{Figure 1}
