% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_cross_topo.R
\name{min_dist}
\alias{min_dist}
\title{Find minimum distance to other species in topology}
\usage{
min_dist(dist, pop, pops)
}
\arguments{
\item{dist}{pair wise distances within topology}

\item{pop}{queried population}

\item{pops}{all populations within the topology}
}
\description{
\code{min_dist} finds the minimum distance of a focal species to all other species in a topology
}
\seealso{
Other Figure 5: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{no_title}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 5}
