% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{network_layout}
\alias{network_layout}
\title{crate a network layout for pairwise species comparisons}
\usage{
network_layout(n, rotate = 0, label = NULL, weight = 2, loc = NA)
}
\arguments{
\item{n}{nomber of nodes}

\item{rotate}{rotation angle}

\item{label}{node labels}

\item{weight}{link weighting}

\item{loc}{location (Belize/ Honduras/ Panama)}
}
\description{
\code{network_layout} creates a network layout for pairwise species comparisons.
}
\seealso{
Other Figure 1: 
\code{\link{make_faint_clr}()},
\code{\link{plot_fish}()},
\code{\link{plot_network}()},
\code{\link{summarize_dxy}()},
\code{\link{summarize_fst}()},
\code{\link{whisker_points}()}
}
\concept{Figure 1}
