% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{no_title}
\alias{no_title}
\title{drop ggplot y axis titles}
\usage{
no_title(...)
}
\description{
\code{no_title} provides further styling for pannels of Figure 3
}
\details{
This function removes the  y axis title and grid elements
of second and third column panels in Figure 3.
}
\seealso{
Other Figure 3: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 3}
