% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{outlr_clr}
\alias{outlr_clr}
\title{Outlier color}
\format{An object of class \code{character} of length 1.}
\usage{
outlr_clr
}
\description{
\code{outlr_clr} defines the color used to highlight Fst outlier windows in plots.
}
\details{
This defines the color used to highlight Fst outlier windows in plots.
}
\keyword{datasets}
