% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF1.R
\name{pca_plot_no_fish}
\alias{pca_plot_no_fish}
\title{PCA plot without fish annotations}
\usage{
pca_plot_no_fish(loc, mode, pc1 = 1, pc2 = 2)
}
\arguments{
\item{loc}{string, three letter location abbreviation}

\item{mode}{string, type of pca data}

\item{pc1}{integer, number of PC for x axis}

\item{pc2}{integer, number of PC for y axis}
}
\description{
PCA plot without fish annotations
}
