% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{plot_clr}
\alias{plot_clr}
\title{Default plot color}
\format{
An object of class \code{character} of length 1.
}
\usage{
plot_clr
}
\description{
\code{plot_clr} is the default color of data in plots.
}
\details{
This color is used to plot neutral data thoughout the figures.
}
\keyword{datasets}
