% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{plot_fish}
\alias{plot_fish}
\title{Plot single fish}
\usage{
plot_fish(short, x = 0, y = 0, height = 0.75, width = 0.75)
}
\description{
\code{plot_fish} creates a ggplot annotations layer with a single hamlet.
}
\details{
This is used to annotate plots.
}
