% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{plot_fish2}
\alias{plot_fish2}
\title{Plot Hamlet drawing (alt)}
\usage{
plot_fish2(
  name,
  x = 0,
  y = 3,
  height = 4,
  width = 4,
  lwd = 0.15,
  line_color = "black"
)
}
\arguments{
\item{name}{string, species name}

\item{x}{numeric, horizontal center of hamlet image}

\item{y}{numeric, vertical center of hamlet image}

\item{height}{numeric, height of hamlet image}

\item{width}{numeric, width of hamlet image}

\item{lwd}{numeric, line width of hamlet drawing line elements}

\item{line_color}{string, color code, color of hamlet drawing line elements}
}
\description{
Plot hamlet drawing with adjustable line width for line elements of drawing.
}
