% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{plot_fish_lwd}
\alias{plot_fish_lwd}
\title{Hamlet Annotation (alt)}
\usage{
plot_fish_lwd(
  short,
  x = 0,
  y = 3,
  height = 5,
  width = 5,
  lwd = 0.15,
  line_color = "transparent"
)
}
\arguments{
\item{short}{string, three letter species name abbreviation}

\item{x}{numeric, horizontal center of hamlet image}

\item{y}{numeric, vertical center of hamlet image}

\item{height}{numeric, height of hamlet image}

\item{width}{numeric, width of hamlet image}

\item{lwd}{numeric, line width of hamlet drawing line elements}

\item{line_color}{string, color code, color of hamlet drawing line elements}
}
\description{
Add ggplot hamlet drawing  annotation with adjustable line width for line elements of drawing.
}
