% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting_f3.R
\name{plot_fish_zoom}
\alias{plot_fish_zoom}
\title{Plot fish for topology legend}
\usage{
plot_fish_zoom(anno_tab, idx, height = 0.1, width = 5)
}
\description{
\code{plot_fish_zoom} creates a hamlet annotation from  a predefined table.
}
\details{
This funtion is used in the creation of the topologie legend.
}
\seealso{
Other Figure 3: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{no_title}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 3}
