% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{plot_fishes_location}
\alias{plot_fishes_location}
\title{Plot two fishes with location}
\usage{
plot_fishes_location(left, right, loc)
}
\arguments{
\item{left}{string, three letter species identifier}

\item{right}{string, three letter species identifier}

\item{loc}{string, sample location (bel [Belize]/ hon [Honduras]/ pan [Panama])}
}
\description{
\code{plot_fishes_location} creates a ggplot annotation layer with two hamlets and a flag.
}
\details{
This is used to annotate plots.
}
