% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_plotting.R
\name{plot_leg}
\alias{plot_leg}
\title{Create topology legend}
\usage{
plot_leg(
  spec1 = "puella",
  spec2 = "maya",
  color = "#084082ff",
  size = 0.5,
  mode = "pair"
)
}
\arguments{
\item{spec1}{string, three letter species identifier}

\item{spec2}{string, three letter species identifier}

\item{color}{string, color}

\item{size}{numeric, line width}

\item{mode}{string, legend element mode ("pair"/"isolation")}
}
\description{
\code{plot_leg} creates the elements of the topology highlighting legend.
}
\details{
This funtion is used to create the topologie legend of a single a single
highlighing scheme.
Depending on the highlighting mode (pair/ isolation), the background geometry
is created and the hamlet annotations are collected.

Then the geometry is plotted an the hamlets are added.
}
\seealso{
Other Figure 5: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{no_title}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 5}
