% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX10.R
\name{plot_loc}
\alias{plot_loc}
\title{Plot hybridization data}
\usage{
plot_loc(loc)
}
\description{
\code{getPofZ} plot hybridization data of a location
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{collapse_peaks}()},
\code{\link{data_amdx}()},
\code{\link{getPofZ}()},
\code{\link{get_gxp_both_models}()},
\code{\link{reformat_run_name}()}
}
\concept{Figure SX}
