% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{plot_network}
\alias{plot_network}
\title{Plot pariwise species comparisons per location}
\usage{
plot_network(loc, nodes, edges, asp = 0.8, sep = 0,
  node_lab_shift = 0)
}
\description{
\code{plot_network} creates a ggplot with a network of species comparisons.
}
\details{
This creates a network for all pairwise species comparisons within a location.
}
