% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF11.R
\name{plot_outl_tree}
\alias{plot_outl_tree}
\title{plot region tree}
\usage{
plot_outl_tree(tree, show_legend = TRUE)
}
\arguments{
\item{tree}{tree object}

\item{show_legend}{logical}
}
\description{
plot region tree
}
