% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF12.R
\name{plot_outl_tree_s}
\alias{plot_outl_tree_s}
\title{plot region tree (with Serranus outgroup)}
\usage{
plot_outl_tree_s(tree, show_legend = TRUE)
}
\arguments{
\item{tree}{tree object}

\item{show_legend}{logical}
}
\description{
plot region tree (with Serranus outgroup)
}
