% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF5.R
\name{plot_pair_run}
\alias{plot_pair_run}
\title{create hamlet pair annotations}
\usage{
plot_pair_run(run, loc, left, right)
}
\arguments{
\item{run}{string, species pair}

\item{loc}{string, sample location (bel [Belize]/ hon [Honduras]/ pan [Panama])}

\item{left}{string, left species}

\item{right}{string, right species}
}
\description{
\code{plot_pair_run} prepares the hamlet pair annotations.
}
