% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_plotting.R
\name{plot_panel_gxp}
\alias{plot_panel_gxp}
\title{Create the genotype x phenotype panel}
\usage{
plot_panel_gxp(lg, start, end, trait, ...)
}
\arguments{
\item{lg}{string, linkage group (eg. "LG08")}

\item{start}{numeric, start position of window (bp)}

\item{end}{numeric, end position of window (bp)}

\item{trait}{string, trait identifier for icon on gxp plot}

\item{...}{catch-all parameter to allow excessive parameters through purrr::pmap}
}
\description{
\code{plot_panel_gxp} plots the genotype x phenotype panel of a selected outlier window
}
\seealso{
Other Figure 5: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{no_title}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 5}
