% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_plotting.R
\name{plot_panel_twisst}
\alias{plot_panel_twisst}
\title{Create the topology weighting panel}
\usage{
plot_panel_twisst(
  loc = "bel",
  lg,
  start,
  end,
  window_size = twisst_size,
  neighbours,
  xlab = TRUE,
  highlight_mode,
  data_tables,
  ...
)
}
\arguments{
\item{loc}{sample location (bel [Belize]/ hon [Honduras]/ pan [Panama])}

\item{lg}{string, linkage group (eg. "LG08")}

\item{start}{numeric, start position of window (bp)}

\item{end}{numeric, end position of window (bp)}

\item{window_size}{numeric, window size for twisst panels (n SNPS)}

\item{neighbours}{topology highlight grouping}

\item{xlab}{logical, should x axis labels be plotted?}

\item{highlight_mode}{string, topology highlight mode ("pair", "isolation")}

\item{data_tables}{list with twisst results}

\item{...}{catch-all parameter to allow excessive parameters through purrr::pmap}
}
\description{
\code{plot_panel_twisst} plots the topolgy weighting panel of a selected outlier window
}
\details{
The topology weighting data for the selected location and
LG is fetched.
Then the number of possible topolgies is determined based
on the location and all topolgies are imported.
Dependend on th highlighting mode (pair/ isolation),
the topologies of interest are determined.
The topologies of interest are labelled in the dataset and
the color scheme for the highlighting is created.

Finally, the topologie weighting panel is plotted.
}
\seealso{
Other Figure 5: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{no_title}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 5}
