% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_post_review.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot Tree}
\usage{
plot_tree(
  tree,
  angle_in = 0,
  color = "red",
  higl_clade = TRUE,
  higl_node = NA,
  root_node = NA,
  rotate_root = FALSE,
  xlim = c(-2, 4.5),
  support_guide = FALSE,
  open_angle = 180
)
}
\arguments{
\item{tree}{tree object}

\item{angle_in}{numeric, tree opening angle}

\item{color}{string, highlight color}

\item{higl_clade}{logical, should there be highlighting}

\item{higl_node}{node, where to start highlighting}

\item{root_node}{node, where to rotate the tree}

\item{rotate_root}{logical, should the tree be rotated}

\item{xlim}{numeric vector, x limits}

\item{support_guide}{logical, show legend for node support}

\item{open_angle}{numeric, tree opening angle}
}
\description{
Plot Tree
}
