% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{pop_levels}
\alias{pop_levels}
\title{Population order}
\format{
An object of class \code{character} of length 14.
}
\usage{
pop_levels
}
\description{
\code{pop_levels} provides the hierachic order of the populations.
}
\details{
The sampled populatinos are alphabetically ordered, first
by location, then by species.
}
\keyword{datasets}
