% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{process_input}
\alias{process_input}
\title{Process script input}
\usage{
process_input(script_name, args)
}
\description{
\code{process_input} processes the script input and echos the received parameters.
}
