% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{project_case}
\alias{project_case}
\title{Project case}
\usage{
project_case(x)
}
\arguments{
\item{x}{string}
}
\description{
\code{project_case} manages the format of labels for figure panels.
}
\details{
Currently all figure sub panels are labeled using lower case.
This function potentially needs to be changed in case of
resubmission to a different journal.
}
