% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\name{project_inv_case}
\alias{project_inv_case}
\title{Inverted project case}
\usage{
project_inv_case(x)
}
\description{
\code{project_inv_case} manages the format of labels within figures.
}
\details{
Labels within figures (outlier highlights) should be in inverse case
compared to figure panel label to avoid confusion.
This function potentially needs to be changed in case of
resubmission to a different journal.
}
