% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF7.R
\name{read_sc}
\alias{read_sc}
\title{import sc data sheet}
\usage{
read_sc(file)
}
\arguments{
\item{file}{input file}
}
\description{
\code{read_sc} imports a sc data sheet.
}
\seealso{
Other Suppl Figure 7: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{data_amdx}()}
}
\concept{Suppl Figure 7}
