% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{refactor}
\alias{refactor}
\title{Reorder factor levels}
\usage{
refactor(self, globals)
}
\description{
\code{refactor} redorders levels of a factor levels based on an external table.
}
\seealso{
Other General functions: \code{\link{darken}},
  \code{\link{export_2_latex}}, \code{\link{get_dxy}},
  \code{\link{get_fst_fixed}}, \code{\link{get_fst}},
  \code{\link{get_gxp}}, \code{\link{get_pi}},
  \code{\link{join_list}}, \code{\link{lighten}},
  \code{\link{process_input}}
}
\concept{General functions}
