% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF2.R
\name{refactor_run}
\alias{refactor_run}
\title{reorder species pairs}
\usage{
refactor_run(self, globals)
}
\arguments{
\item{self}{string vector, species pairs}

\item{globals}{reference table, order of species pairs according to reference table}
}
\description{
\code{refactor_run} reorder species pairs according to fst.
}
\seealso{
Other Suppl Figure 2: 
\code{\link{anno_pair_flag}()},
\code{\link{fst_bar_row_run}()},
\code{\link{plot_pair_run}()},
\code{\link{rescale_fst}()}
}
\concept{Suppl Figure 2}
