% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{reformat_run_name}
\alias{reformat_run_name}
\title{Reformat run name}
\usage{
reformat_run_name(run)
}
\arguments{
\item{run}{string, species pair}
}
\description{
\code{reformat_run_name} formats run to "specloc-specloc" format.
}
