% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF5.R
\name{rescale_fst}
\alias{rescale_fst}
\title{normalize global fst}
\usage{
rescale_fst(fst)
}
\arguments{
\item{fst}{numeric, global fst value}
}
\description{
\code{rescale_fst} normalize global fst to 0 - 1.
}
