% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f6.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Custom reverse log transformation}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{}
}
\description{
Custom reverse log transformation
}
