% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_post_review.R
\name{root_hamlets}
\alias{root_hamlets}
\title{Root hamlet tree}
\usage{
root_hamlets(tree, type)
}
\arguments{
\item{tree}{string, input tree file}

\item{type}{string, tree type}
}
\description{
Just for the flexible outgroup.
}
