% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF6.R
\name{root_manual}
\alias{root_manual}
\title{reroot tree manually}
\usage{
root_manual(object, outgroup, node = NULL, resolve.root = TRUE, ...)
}
\description{
\code{root_manual} reroot a tree manually.
}
\seealso{
Other Suppl Figure 6: 
\code{\link{StatTreeData}},
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{setup_tree_data}()}
}
\concept{Suppl Figure 6}
