% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{shps}
\alias{shps}
\title{Location shapes}
\format{An object of class \code{numeric} of length 4.}
\usage{
shps
}
\description{
\code{shps} defines the shapes for sample locations within plots.
}
\details{
This is used in shape scales thoughout the paper
where the locations are indicated by shape.
}
\keyword{datasets}
