% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{sp_labs}
\alias{sp_labs}
\title{Species labels}
\format{
An object of class \code{expression} of length 11.
}
\usage{
sp_labs
}
\description{
\code{sp_labs} stores the scientific names of the hamlet abrreviations.
}
\details{
This is used to translate between the 3-letter species abrreviations and
the actual formated scientific names of the hamlet species.
}
\keyword{datasets}
