% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{sp_names}
\alias{sp_names}
\title{Species names}
\format{
An object of class \code{character} of length 11.
}
\usage{
sp_names
}
\description{
\code{sp_names} stores the species names of the hamlet abrreviations.
}
\details{
This is used to translate between the 3-letter species abrreviations and
the actual species names of the hamlet species.
}
\keyword{datasets}
