% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF6.R
\docType{data}
\name{StatTreeData}
\alias{StatTreeData}
\title{ggplot layer for node support}
\format{
An object of class \code{StatTreeLabel} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTreeData
}
\description{
\code{StatTreeData} provides a custom ggplot layer for node support.
}
\seealso{
Other Suppl Figure 6: 
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()}
}
\concept{Suppl Figure 6}
\keyword{datasets}
