% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{summarize_dxy}
\alias{summarize_dxy}
\title{Import and summarize dxy data}
\usage{
summarize_dxy(file)
}
\description{
\code{summarize_dxy} imports and summarizes dxy data.
}
\details{
dxy data is imported, filtered for non-overlapping windows
and boxplot statistics are created.
}
