% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{summarize_fst}
\alias{summarize_fst}
\title{Import and summarize Fst data}
\usage{
summarize_fst(file)
}
\description{
\code{summarize_fst} imports and summarizes Fst data.
}
\details{
Fst data is imported, filtered for non-overlapping windows
and boxplot statistics are created.
}
