% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f6.R
\name{theme_gradient_bg_x}
\alias{theme_gradient_bg_x}
\title{Tiled Grob background for ggplot}
\usage{
theme_gradient_bg_x(
  xlim_in,
  colors = c("#7B0664", "#E32219"),
  fun = identity,
  step = 1,
  ...
)
}
\arguments{
\item{xlim_in}{numeric vector, x range of ggplot}

\item{colors}{string vector, colors to use for background gradients}

\item{fun}{function, x transformation function}

\item{step}{numeric, step width for x breaks (after transformation)}

\item{...}{catch all parameter, for usage with purrr::pmap() and unexpected parameters}
}
\description{
Tiled Grob background for ggplot
}
