% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF4.R
\name{theme_hyb}
\alias{theme_hyb}
\title{Custom ggplot theme}
\usage{
theme_hyb(legend.position = "none", ...)
}
\arguments{
\item{legend.position}{string, according to ggplot2::theme()}

\item{...}{arguments funneled through to ggplot2::theme()}
}
\description{
\code{theme_hyb} ggplot theme for hybridization plots
}
