% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{whisker_points}
\alias{whisker_points}
\title{Whisker points}
\usage{
whisker_points(x)
}
\description{
\code{whisker_points} calculates the extent of boxplot whiskers.
}
\details{
This is used to plot boxplots without outliers.
}
