% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF13.R
\name{add_spec_drawing}
\alias{add_spec_drawing}
\title{Add fish annotation}
\usage{
add_spec_drawing(spec, pos)
}
\arguments{
\item{spec}{hamlet species}

\item{pos}{position on ggplot}
}
\description{
\code{add_spec_drawing} adds a fish annotation to a plot
}
