% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX9.R
\name{add_spec_drawing}
\alias{add_spec_drawing}
\title{Add fish annotation}
\usage{
add_spec_drawing(spec, pos)
}
\description{
\code{add_spec_drawing} adds a fish annotation to a plot
}
\seealso{
Other Figure SX: 
\code{\link{adm_plot}()},
\code{\link{collapse_peaks}()},
\code{\link{data_amdx}()},
\code{\link{getPofZ}()},
\code{\link{get_gxp_both_models}()},
\code{\link{label_spacer}()},
\code{\link{plot_loc}()},
\code{\link{reformat_run_name}()},
\code{\link{theme_hyb}()}
}
\concept{Figure SX}
