% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF13.R
\name{adm_plot}
\alias{adm_plot}
\title{Plot adimxture data}
\usage{
adm_plot(data, gid_in)
}
\arguments{
\item{data}{tibble, input data}

\item{gid_in}{string, identifier of fst outlier ID (eg "LG04_1")}
}
\description{
\code{adm_plot} plots the admixture data
}
