% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF7.R
\name{adm_plot}
\alias{adm_plot}
\title{Plot adimxture data}
\usage{
adm_plot(gid_in)
}
\arguments{
\item{gid_in}{string, identifier of fst outlier ID (eg "LG04_1")}
}
\description{
\code{adm_plot} plots the admixture data
}
\seealso{
Other Suppl Figure 7: 
\code{\link{add_spec_drawing}()},
\code{\link{data_amdx}()},
\code{\link{read_sc}()}
}
\concept{Suppl Figure 7}
