% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF2.R
\name{anno_pair_flag}
\alias{anno_pair_flag}
\title{create hamlet pair annotations}
\usage{
anno_pair_flag(
  loc,
  left,
  right,
  circle_color = NA,
  circle_fill_left = "white",
  circle_fill_right = "lightgray",
  circle_lwd = 0.5,
  plot_names = FALSE,
  plot_name_size = 3,
  font_family = "sans",
  ...
)
}
\arguments{
\item{loc}{string, sample location (bel [Belize]/ hon [Honduras]/ pan [Panama])}

\item{left}{string, left species}

\item{right}{string, right species}

\item{circle_color}{string, color}

\item{circle_fill_left}{string, color}

\item{circle_fill_right}{string, color}

\item{circle_lwd}{numeric}

\item{plot_names}{logical, should species labels be plotted}

\item{plot_name_size}{numeric, size of labels}

\item{font_family}{string, font family}

\item{...}{catch all for excess parameters for purrr::pmap}
}
\description{
\code{plot_pair_run} creates the hamlet pair annotations.
}
\seealso{
Other Suppl Figure 2: 
\code{\link{fst_bar_row_run}()},
\code{\link{plot_pair_run}()},
\code{\link{refactor_run}()},
\code{\link{rescale_fst}()}
}
\concept{Suppl Figure 2}
