% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX8.R
\name{collapse_peaks}
\alias{collapse_peaks}
\title{Collapse all Fst peaks of a given run}
\usage{
collapse_peaks(x)
}
\description{
\code{collapse_peaks} collapses all Fst peaks of a given run
}
\seealso{
Other Figure SX: 
\code{\link{reformat_run_name}()}
}
\concept{Figure SX}
