% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF3.R
\name{collapse_peaks}
\alias{collapse_peaks}
\title{Collapse all Fst peaks of a given run}
\usage{
collapse_peaks(x)
}
\arguments{
\item{x}{string vector with outlier peak ids}
}
\description{
\code{collapse_peaks} collapses all Fst peaks of a given run
}
\seealso{
Other Suppl Figure 3: 
\code{\link{geom_hypo_grob2}()},
\code{\link{get_fst_fixed}()},
\code{\link{plot_fishes_location}()},
\code{\link{reformat_run_name}()},
\code{\link{summarise_model}()}
}
\concept{Suppl Figure 3}
