% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX8.R
\name{collapse_peaks}
\alias{collapse_peaks}
\title{Collapse all Fst peaks of a given run}
\usage{
collapse_peaks(x)
}
\description{
\code{collapse_peaks} collapses all Fst peaks of a given run
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{data_amdx}()},
\code{\link{getPofZ}()},
\code{\link{get_gxp_both_models}()},
\code{\link{label_spacer}()},
\code{\link{plot_loc}()},
\code{\link{reformat_run_name}()},
\code{\link{theme_hyb}()}
}
\concept{Figure SX}
