% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_plotting.R
\name{custom_annoplot}
\alias{custom_annoplot}
\title{Custom annotation base plot}
\usage{
custom_annoplot(
  ...,
  searchLG,
  xrange,
  genes_of_interest = c(),
  genes_of_sec_interest = c(),
  anno_rown = 3,
  width = 0.1,
  gene_color = "darkgray",
  start,
  end
)
}
\arguments{
\item{...}{catch-all parameter to allow excessive parameters through purrr::pmap}

\item{searchLG}{string, linkage group (eg. "LG08")}

\item{xrange}{numeric vector of length 2, range of the plotted window}

\item{genes_of_interest}{vector of strings, genes that receive a label}

\item{genes_of_sec_interest}{vector of strings (deprecate/legacy)}

\item{anno_rown}{numeric, number of rows used for gene annotation}

\item{width}{numeric, width of genes}

\item{gene_color}{string (color), highlight color for genes}

\item{start}{numeric, start position of window (bp)}

\item{end}{numeric, end position of window (bp)}
}
\description{
\code{custom_annoplot} extracts and plots hypoplectrus annotation data
}
\details{
This is a modification of hypogen::hypo_annotation_baseplot().
The highlighting of the outlier area is added and the coloration
of genes is simplified to ony a single color.
}
\seealso{
Other Figure 5: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{no_title}()},
\code{\link{plot_curtain}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 5}
