% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX9.R
\name{data_amdx}
\alias{data_amdx}
\title{Get adimxture data}
\usage{
data_amdx(gid, k, admx_path)
}
\description{
\code{data_amdx} imports the admixture data
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{collapse_peaks}()},
\code{\link{getPofZ}()},
\code{\link{plot_loc}()},
\code{\link{reformat_run_name}()}
}
\concept{Figure SX}
