% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_config.R
\docType{data}
\name{f_width}
\alias{f_width}
\title{Default figure width}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
f_width
}
\description{
\code{f_width} defines the full width for figures (mm).
}
\details{
7 inch = 178m
}
\keyword{datasets}
