% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX4.R
\name{fish_plot}
\alias{fish_plot}
\title{hamlet pair annotation}
\usage{
fish_plot(spec)
}
\description{
\code{fish_plot} provides hamlet pair annotation.
}
\seealso{
Other Suppl. Figure SX: 
\code{\link{StatTreeData}},
\code{\link{anno_pair_flag}()},
\code{\link{fish_plot2}()},
\code{\link{fst_bar_row_run}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{plot_fishes_location}()},
\code{\link{plot_pair_run}()},
\code{\link{refactor_run}()},
\code{\link{rescale_fst}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()},
\code{\link{summarise_model}()}
}
\concept{Suppl. Figure SX}
