% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF5.R
\name{fish_plot}
\alias{fish_plot}
\title{hamlet pair annotation}
\usage{
fish_plot(spec)
}
\description{
\code{fish_plot} provides hamlet pair annotation.
}
\seealso{
Other Suppl Figure 5: 
\code{\link{fish_plot2}()}
}
\concept{Suppl Figure 5}
