% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF10.R
\name{fish_plot2}
\alias{fish_plot2}
\title{hamlet pair annotation 2}
\usage{
fish_plot2(spec)
}
\arguments{
\item{spec}{string six letter population name}
}
\description{
\code{fish_plot2} provides hamlet pair annotation with alternative axis specifications.
}
