% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF5.R
\name{fish_plot2}
\alias{fish_plot2}
\title{hamlet pair annotation 2}
\usage{
fish_plot2(spec)
}
\description{
\code{fish_plot2} provides hamlet pair annotation with alternative axis specifications.
}
\seealso{
Other Suppl Figure 5: 
\code{\link{fish_plot}()}
}
\concept{Suppl Figure 5}
