% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF2.R
\name{fst_bar_row_run}
\alias{fst_bar_row_run}
\title{crate global fst bar table}
\usage{
fst_bar_row_run(fst, run)
}
\description{
\code{fst_bar_row_run} crates a rescaled global fst bar table.
}
\seealso{
Other Suppl Figure 2: 
\code{\link{anno_pair_flag}()},
\code{\link{plot_pair_run}()},
\code{\link{refactor_run}()},
\code{\link{rescale_fst}()}
}
\concept{Suppl Figure 2}
