% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF5.R
\name{fst_bar_row_run}
\alias{fst_bar_row_run}
\title{crate global fst bar table}
\usage{
fst_bar_row_run(fst, run)
}
\arguments{
\item{fst}{global fst value}

\item{run}{string, species pair}
}
\description{
\code{fst_bar_row_run} crates a re-scaled global fst bar table.
}
