% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{geom_hypo_grob2}
\alias{geom_hypo_grob2}
\title{The custom grob geom}
\usage{
geom_hypo_grob2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping; ggplot2::aes()}

\item{data}{data frame}

\item{stat}{ggplot2 stat}

\item{position}{string, "identity"}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{...}{arguments passed to ggplot2::layer()}
}
\description{
This geom builds heavily on the answer by baptiste on the
tidiverse github forum:
https://github.com/tidyverse/ggplot2/issues/1399
}
