% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF6.R
\name{geom_nodepoint_support}
\alias{geom_nodepoint_support}
\title{backbone for ggplot layer for node support}
\usage{
geom_nodepoint_support(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping; ggplot2::aes()}

\item{data}{data frame}

\item{position}{string, "identity"}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{...}{arguments passed to geom_point_support()}
}
\description{
\code{geom_nodepoint_support} provides the backbone for a custom ggplot layer for node support.
}
\seealso{
Other Suppl Figure 6: 
\code{\link{StatTreeData}},
\code{\link{get_tree}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()}
}
\concept{Suppl Figure 6}
