% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX7.R
\name{geom_nodepoint_support}
\alias{geom_nodepoint_support}
\title{backbone for ggplot layer for node support}
\usage{
geom_nodepoint_support(mapping = NULL, data = NULL,
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
\code{geom_nodepoint_support} provides the backbone for a custom ggplot layer for node support.
}
\seealso{
Other Suppl. Figure SX: \code{\link{StatTreeData}},
  \code{\link{anno_pair_flag}}, \code{\link{fish_plot2}},
  \code{\link{fish_plot}}, \code{\link{fst_bar_row_run}},
  \code{\link{geom_hypo_grob2}}, \code{\link{get_tree}},
  \code{\link{plot_fishes_location}},
  \code{\link{plot_pair_run}}, \code{\link{refactor_run}},
  \code{\link{rescale_fst}}, \code{\link{root_manual}},
  \code{\link{setup_tree_data}},
  \code{\link{summarise_model}}
}
\concept{Suppl. Figure SX}
