% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f4.R
\name{get_cc}
\alias{get_cc}
\title{Import cross coalescence data}
\usage{
get_cc(file, cc_groups, cc_path, mu = 3.7e-08, gen = 1)
}
\description{
\code{get_msmc} imports cross coalescence data and extracts according run information.
}
\details{
The data is imported and run components are extracted from the a reference table,
and added to the data.
}
\seealso{
Other Figure 4: \code{\link{get_msmc}}
}
\concept{Figure 4}
