% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{get_dxy}
\alias{get_dxy}
\title{Import dxy data}
\usage{
get_dxy(file, kb = "50k")
}
\arguments{
\item{file}{input file}

\item{kb}{string, window size ("10k"/"50k")}
}
\description{
\code{get_dxy} imports dxy data, adds the genomic position and the run.
}
\details{
The run name (the pair wise species comparison) is extracted from the file name,
the data is imported, column names are standardized and genomic position and
run name are added.
}
\seealso{
Other General functions: 
\code{\link{darken}()},
\code{\link{export_2_latex}()},
\code{\link{get_fst_summary_data}()},
\code{\link{get_fst}()},
\code{\link{get_gxp}()},
\code{\link{get_pi}()},
\code{\link{join_list}()},
\code{\link{lighten}()},
\code{\link{process_input}()},
\code{\link{refactor}()}
}
\concept{General functions}
