% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX8.R
\name{get_fst_fixed}
\alias{get_fst_fixed}
\title{Compute summary stats an absolute for Fst outlier}
\usage{
get_fst_fixed(file, run, fst_threshold, ...)
}
\description{
\code{get_fst_percentile} computes summary stats for Fst outlier for a given Fst threshold.
}
\seealso{
Other General functions: \code{\link{darken}},
  \code{\link{export_2_latex}}, \code{\link{get_dxy}},
  \code{\link{get_fst}}, \code{\link{get_gxp}},
  \code{\link{get_pi}}, \code{\link{join_list}},
  \code{\link{lighten}}, \code{\link{process_input}},
  \code{\link{refactor}}
}
\concept{General functions}
