% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF3.R
\name{get_fst_fixed}
\alias{get_fst_fixed}
\title{Compute summary stats an absolute for Fst outlier}
\usage{
get_fst_fixed(file, run, fst_threshold, ...)
}
\description{
\code{get_fst_percentile} computes summary stats for Fst outlier for a given Fst threshold.
}
\seealso{
Other Suppl Figure 3: 
\code{\link{collapse_peaks}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{plot_fishes_location}()},
\code{\link{reformat_run_name}()},
\code{\link{summarise_model}()}
}
\concept{Suppl Figure 3}
