% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{get_fst_fixed}
\alias{get_fst_fixed}
\title{Compute summary stats an absolute for Fst outlier}
\usage{
get_fst_fixed(file, run, fst_threshold, ...)
}
\arguments{
\item{file}{path to fst file}

\item{run}{string, species pair}

\item{fst_threshold}{numeric, threshold for fst outliers}

\item{...}{arguments passed to hypogen::hypo_import_windows()}
}
\description{
\code{get_fst_percentile} computes summary stats for Fst outlier for a given Fst threshold.
}
