% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tables.R
\name{get_genes}
\alias{get_genes}
\title{Collect hamlet genes}
\usage{
get_genes(gid, chrom, start, end, ...)
}
\description{
\code{get_genes} collects the genes within a outlier window.
}
\seealso{
Other Table functions: 
\code{\link{as_multirow}()}
}
\concept{Table functions}
