% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF9.R
\name{get_gxp_both_models}
\alias{get_gxp_both_models}
\title{Get gxp data}
\usage{
get_gxp_both_models(file, trait, model_type, path)
}
\description{
\code{get_gxp_both_models} imports gxp data
}
\concept{Suppl Figure 9}
