% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF4.R
\name{getPofZ}
\alias{getPofZ}
\title{Get hybridization data}
\usage{
getPofZ(base_dir, folder)
}
\arguments{
\item{base_dir}{path containing the newhybrids results}

\item{folder}{sub-folder of interest within the results folder}
}
\description{
\code{getPofZ} imports hybridization data
}
