% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX10.R
\name{getPofZ}
\alias{getPofZ}
\title{Get hybridization data}
\usage{
getPofZ(base_dir, folder)
}
\description{
\code{getPofZ} imports hybridization data
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{collapse_peaks}()},
\code{\link{data_amdx}()},
\code{\link{get_gxp_both_models}()},
\code{\link{plot_loc}()},
\code{\link{reformat_run_name}()}
}
\concept{Figure SX}
