% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{join_list}
\alias{join_list}
\title{Iterate left_join over list}
\usage{
join_list(lst)
}
\description{
\code{join_list} binds a list of tables according to a common column.
}
\seealso{
Other General functions: 
\code{\link{darken}()},
\code{\link{export_2_latex}()},
\code{\link{get_dxy}()},
\code{\link{get_fst_fixed}()},
\code{\link{get_fst_summary_data}()},
\code{\link{get_fst}()},
\code{\link{get_gxp}()},
\code{\link{get_pi}()},
\code{\link{lighten}()},
\code{\link{process_input}()},
\code{\link{refactor}()}
}
\concept{General functions}
