% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX10.R
\name{label_spacer}
\alias{label_spacer}
\title{Adjust plot sizes}
\usage{
label_spacer(x, plus = 1.1)
}
\description{
\code{label_spacer} adjusts plot sizes with respect to presence of labels
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{collapse_peaks}()},
\code{\link{data_amdx}()},
\code{\link{getPofZ}()},
\code{\link{get_gxp_both_models}()},
\code{\link{plot_loc}()},
\code{\link{reformat_run_name}()},
\code{\link{theme_hyb}()}
}
\concept{Figure SX}
