% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF1.R
\name{label_spacer}
\alias{label_spacer}
\title{Adjust plot sizes}
\usage{
label_spacer(x, plus = 1.1)
}
\arguments{
\item{x}{label position}

\item{plus}{label offset}
}
\description{
\code{label_spacer} adjusts plot sizes with respect to presence of labels
}
\seealso{
Other Suppl Figure 1: 
\code{\link{getPofZ}()},
\code{\link{plot_loc}()},
\code{\link{theme_hyb}()}
}
\concept{Suppl Figure 1}
