% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f4.R
\name{match_twisst_files}
\alias{match_twisst_files}
\title{Import all topology weighting data sets for a location}
\usage{
match_twisst_files(loc, window_size = 200, panel = "a")
}
\arguments{
\item{loc}{sample location (bel [Belize]/ hon [Honduras]/ pan [Panama])}

\item{window_size}{window size of twisst run}

\item{panel}{figure panel identifier (for labeling)}
}
\description{
\code{match_twisst_files} imports all topology weighting data sets for a location.
}
\details{
First the weight and position files of all topology weighting data
sets (the different linkage groups) for a given location are collected.

The topology weighting data import function is iterated over all
data sets and combined to create a full data set.

Then, the full data set is summarized to then be able to rank
the typologies based on their overall weighting.

The topology ranks are added to the full data set and a window
column containing the location is added for faceting.
}
\seealso{
Other Figure 4: 
\code{\link{get_twisst_data}()}
}
\concept{Figure 4}
