% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{network_layout}
\alias{network_layout}
\title{crate a network layout for pairwise species comparisons}
\usage{
network_layout(n, rotate = 0, label = NULL, weight = 2, loc = NA)
}
\description{
\code{network_layout} creates a network layout for pairwise species comparisons.
}
