% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f5_plotting.R
\name{plot_curtain}
\alias{plot_curtain}
\title{Manage plot panels for a single outlier window}
\usage{
plot_curtain(
  loc = "bel",
  outlier_id,
  outlier_nr,
  lg,
  start,
  end,
  cool_genes,
  text = TRUE,
  label,
  trait,
  data_tables,
  ...
)
}
\arguments{
\item{loc}{string, sample location (bel [Belize]/ hon [Honduras]/ pan [Panama])}

\item{outlier_id}{string, identifier of fst outlier ID (eg "LG04_1")}

\item{outlier_nr}{numeric (deprecated)}

\item{lg}{string, linkage group (eg. "LG08")}

\item{start}{numeric, start position of window (bp)}

\item{end}{numeric, end position of window (bp)}

\item{cool_genes}{vector of strings, genes that receive a label}

\item{text}{logical, toggle y axis label and ticks}

\item{label}{string, panel label ("A"/"B"/"C")}

\item{trait}{string, trait identifier for icon on gxp plot}

\item{data_tables}{list with twisst results}

\item{...}{catch-all parameter to allow excessive parameters through purrr::pmap}
}
\description{
\code{plot_curtain} calls all individual plotting functions needed for a single outlier zoom
}
\details{
The panels containg the gene models, fst, dxy delta dxy,
genotype x phenotype association and topology weighting are called.

Then they are combined using \code{cowplot::plot_grid} and
depending on the column position (first column/ other column),
the y axis titles are removed.
}
\seealso{
Other Figure 5: 
\code{\link{ax_scl}()},
\code{\link{cross_spec}()},
\code{\link{custom_annoplot}()},
\code{\link{dist_tibble}()},
\code{\link{distances_tree}()},
\code{\link{get_clr}()},
\code{\link{get_dist}()},
\code{\link{get_gxp_long}()},
\code{\link{get_isolated_topos}()},
\code{\link{get_neighbour_topos}()},
\code{\link{is_isolated}()},
\code{\link{min_dist}()},
\code{\link{no_title}()},
\code{\link{plot_fish_zoom}()},
\code{\link{plot_fst_poptree}()},
\code{\link{plot_leg}()},
\code{\link{plot_panel_anno}()},
\code{\link{plot_panel_delta_dxy}()},
\code{\link{plot_panel_dxy}()},
\code{\link{plot_panel_fst}()},
\code{\link{plot_panel_gxp}()},
\code{\link{plot_panel_twisst}()},
\code{\link{prep_data}()},
\code{\link{theme_panels}()}
}
\concept{Figure 5}
