% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF3.R
\name{plot_fishes_location}
\alias{plot_fishes_location}
\title{Plot two fishes with location}
\usage{
plot_fishes_location(left, right, loc)
}
\description{
\code{plot_fishes_location} creates a ggplot annotation layer with two hamlets and a flag.
}
\details{
This is used to annotate plots.
}
\seealso{
Other Suppl Figure 3: 
\code{\link{collapse_peaks}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{get_fst_fixed}()},
\code{\link{reformat_run_name}()},
\code{\link{summarise_model}()}
}
\concept{Suppl Figure 3}
