% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX1.R
\name{plot_fishes_location}
\alias{plot_fishes_location}
\title{Plot two fishes with location}
\usage{
plot_fishes_location(left, right, loc)
}
\description{
\code{plot_fishes_location} creates a ggplot annotation layer with two hamlets and a flag.
}
\details{
This is used to annotate plots.
}
\seealso{
Other Suppl. Figure SX: \code{\link{geom_hypo_grob2}},
  \code{\link{summarise_model}}
}
\concept{Suppl. Figure SX}
