% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF1.R
\name{plot_loc}
\alias{plot_loc}
\title{Plot hybridization data}
\usage{
plot_loc(loc)
}
\description{
\code{getPofZ} plot hybridization data of a location
}
\seealso{
Other Suppl Figure 1: 
\code{\link{getPofZ}()},
\code{\link{label_spacer}()},
\code{\link{theme_hyb}()}
}
\concept{Suppl Figure 1}
