% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF4.R
\name{plot_loc}
\alias{plot_loc}
\title{Plot hybridization data}
\usage{
plot_loc(loc)
}
\arguments{
\item{loc}{sample location (bel [Belize]/ hon [Honduras]/ pan [Panama])}
}
\description{
\code{getPofZ} plot hybridization data of a location
}
