% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF2.R
\name{plot_pair_run}
\alias{plot_pair_run}
\title{create hamlet pair annotations}
\usage{
plot_pair_run(run, loc, left, right)
}
\description{
\code{plot_pair_run} prepares the hamlet pair annotations.
}
\seealso{
Other Suppl Figure 2: 
\code{\link{anno_pair_flag}()},
\code{\link{fst_bar_row_run}()},
\code{\link{refactor_run}()},
\code{\link{rescale_fst}()}
}
\concept{Suppl Figure 2}
