% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX3.R
\name{plot_pair_run}
\alias{plot_pair_run}
\title{create hamlet pair annotations}
\usage{
plot_pair_run(run, loc, left, right)
}
\description{
\code{plot_pair_run} prepares the hamlet pair annotations.
}
\seealso{
Other Suppl. Figure SX: 
\code{\link{StatTreeData}},
\code{\link{anno_pair_flag}()},
\code{\link{fish_plot2}()},
\code{\link{fish_plot}()},
\code{\link{fst_bar_row_run}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{plot_fishes_location}()},
\code{\link{refactor_run}()},
\code{\link{rescale_fst}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()},
\code{\link{summarise_model}()}
}
\concept{Suppl. Figure SX}
