% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{prep_data}
\alias{prep_data}
\title{Prepate topology weighting data for zoom plot}
\usage{
prep_data(loc, window_size = twisst_size, ...)
}
\description{
\code{prep_data} imports the twisst data of a selected LG
}
\details{
Based on the focal outlier names, the linkage group is determined.
The topology weighting results and positions files are located and
filtered for the focal LG and location.

Then, the selected files are imported.
}
\seealso{
Other Figure 3: \code{\link{ax_scl}},
  \code{\link{cross_spec}}, \code{\link{custom_annoplot}},
  \code{\link{dist_tibble}}, \code{\link{distances_tree}},
  \code{\link{get_clr}}, \code{\link{get_dist}},
  \code{\link{get_gxp_long}},
  \code{\link{get_isolated_topos}},
  \code{\link{get_neighbour_topos}},
  \code{\link{is_isolated}}, \code{\link{min_dist}},
  \code{\link{no_title}}, \code{\link{plot_curtain}},
  \code{\link{plot_fish_zoom}}, \code{\link{plot_leg}},
  \code{\link{plot_panel_anno}},
  \code{\link{plot_panel_delta_dxy}},
  \code{\link{plot_panel_dxy}},
  \code{\link{plot_panel_fst}},
  \code{\link{plot_panel_gxp}},
  \code{\link{plot_panel_twisst}},
  \code{\link{theme_panels}}
}
\concept{Figure 3}
