% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_general.R
\name{process_input}
\alias{process_input}
\title{Process script input}
\usage{
process_input(script_name, args)
}
\description{
\code{process_input} processes the script input and echos the received parameters.
}
\seealso{
Other General functions: \code{\link{darken}},
  \code{\link{export_2_latex}}, \code{\link{get_dxy}},
  \code{\link{get_fst_fixed}}, \code{\link{get_fst}},
  \code{\link{get_gxp}}, \code{\link{get_pi}},
  \code{\link{join_list}}, \code{\link{lighten}},
  \code{\link{refactor}}
}
\concept{General functions}
