% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF5.R
\name{refactor_run}
\alias{refactor_run}
\title{reorder species pairs}
\usage{
refactor_run(self, globals)
}
\arguments{
\item{self}{string vector, species pairs}

\item{globals}{reference table, order of species pairs according to reference table}
}
\description{
\code{refactor_run} reorder species pairs according to fst.
}
