% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF3.R
\name{reformat_run_name}
\alias{reformat_run_name}
\title{Reformat run name}
\usage{
reformat_run_name(run)
}
\arguments{
\item{run}{string, species pair}
}
\description{
\code{reformat_run_name} formats run to "specloc-specloc" format.
}
\seealso{
Other Suppl Figure 3: 
\code{\link{collapse_peaks}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{get_fst_fixed}()},
\code{\link{plot_fishes_location}()},
\code{\link{summarise_model}()}
}
\concept{Suppl Figure 3}
