% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX8.R
\name{reformat_run_name}
\alias{reformat_run_name}
\title{Reformat run name}
\usage{
reformat_run_name(run)
}
\description{
\code{reformat_run_name} formats run to "specloc-specloc" format.
}
\seealso{
Other Figure SX: \code{\link{collapse_peaks}}
}
\concept{Figure SX}
