% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX8.R
\name{reformat_run_name}
\alias{reformat_run_name}
\title{Reformat run name}
\usage{
reformat_run_name(run)
}
\description{
\code{reformat_run_name} formats run to "specloc-specloc" format.
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{collapse_peaks}()},
\code{\link{data_amdx}()},
\code{\link{getPofZ}()},
\code{\link{get_gxp_both_models}()},
\code{\link{plot_loc}()}
}
\concept{Figure SX}
