% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX3.R
\name{rescale_fst}
\alias{rescale_fst}
\title{normalize global fst}
\usage{
rescale_fst(fst)
}
\description{
\code{rescale_fst} normalize global fst to 0 - 1.
}
\seealso{
Other Suppl. Figure SX: 
\code{\link{StatTreeData}},
\code{\link{anno_pair_flag}()},
\code{\link{fish_plot2}()},
\code{\link{fish_plot}()},
\code{\link{fst_bar_row_run}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{plot_fishes_location}()},
\code{\link{plot_pair_run}()},
\code{\link{refactor_run}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()},
\code{\link{summarise_model}()}
}
\concept{Suppl. Figure SX}
