% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF2.R
\name{rescale_fst}
\alias{rescale_fst}
\title{normalize global fst}
\usage{
rescale_fst(fst)
}
\description{
\code{rescale_fst} normalize global fst to 0 - 1.
}
\seealso{
Other Suppl Figure 2: 
\code{\link{anno_pair_flag}()},
\code{\link{fst_bar_row_run}()},
\code{\link{plot_pair_run}()},
\code{\link{refactor_run}()}
}
\concept{Suppl Figure 2}
