% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF6.R
\name{root_manual}
\alias{root_manual}
\title{reroot tree manually}
\usage{
root_manual(object, outgroup, node = NULL, resolve.root = TRUE, ...)
}
\arguments{
\item{object}{tree object}

\item{outgroup}{vector of nodes to define outgroup}

\item{node}{node to root tree at}

\item{resolve.root}{logical}

\item{...}{catch-all parameter to allow excessive parameters through purrr::pmap}
}
\description{
\code{root_manual} reroot a tree manually.
}
\seealso{
Other Suppl Figure 6: 
\code{\link{StatTreeData}},
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{setup_tree_data}()}
}
\concept{Suppl Figure 6}
