% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF6.R
\name{setup_tree_data}
\alias{setup_tree_data}
\title{setup phylo tree data}
\usage{
setup_tree_data(data)
}
\description{
\code{setup_tree_data} sets up phylo tree data.
}
\seealso{
Other Suppl Figure 6: 
\code{\link{StatTreeData}},
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{root_manual}()}
}
\concept{Suppl Figure 6}
