% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX7.R
\docType{data}
\name{StatTreeData}
\alias{StatTreeData}
\title{ggplot layer for node support}
\format{
An object of class \code{StatTreeLabel} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTreeData
}
\description{
\code{StatTreeData} provides a custom ggplot layer for node support.
}
\seealso{
Other Suppl. Figure SX: 
\code{\link{anno_pair_flag}()},
\code{\link{fish_plot2}()},
\code{\link{fish_plot}()},
\code{\link{fst_bar_row_run}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{geom_nodepoint_support}()},
\code{\link{get_tree}()},
\code{\link{plot_fishes_location}()},
\code{\link{plot_pair_run}()},
\code{\link{refactor_run}()},
\code{\link{rescale_fst}()},
\code{\link{root_manual}()},
\code{\link{setup_tree_data}()},
\code{\link{summarise_model}()}
}
\concept{Suppl. Figure SX}
\keyword{datasets}
