% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF3.R
\name{summarise_model}
\alias{summarise_model}
\title{summarize model statistics}
\usage{
summarise_model(data)
}
\arguments{
\item{data}{tibble with column 'mod', containing models to be summarized}
}
\description{
\code{summarise_model} summarizes model statistics in a tibble row.
}
\seealso{
Other Suppl Figure 3: 
\code{\link{collapse_peaks}()},
\code{\link{geom_hypo_grob2}()},
\code{\link{get_fst_fixed}()},
\code{\link{plot_fishes_location}()},
\code{\link{reformat_run_name}()}
}
\concept{Suppl Figure 3}
