% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f3.R
\name{summarise_model}
\alias{summarise_model}
\title{summarize model statistics}
\usage{
summarise_model(data)
}
\arguments{
\item{data}{tibble with column 'mod', containing models to be summarized}
}
\description{
\code{summarise_model} summarizes model statistics in a tibble row.
}
