% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX1.R
\name{summarise_model}
\alias{summarise_model}
\title{summarise model statistics}
\usage{
summarise_model(data)
}
\description{
\code{summarise_model} sumamrizes model statistics in a tibble row.
}
\seealso{
Other Suppl. Figure SX: \code{\link{geom_hypo_grob2}},
  \code{\link{plot_fishes_location}}
}
\concept{Suppl. Figure SX}
