% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SX10.R
\name{theme_hyb}
\alias{theme_hyb}
\title{Custom ggplot theme}
\usage{
theme_hyb(legend.position = "none", ...)
}
\description{
\code{theme_hyb} ggplot theme for hybridization plots
}
\seealso{
Other Figure SX: 
\code{\link{add_spec_drawing}()},
\code{\link{adm_plot}()},
\code{\link{collapse_peaks}()},
\code{\link{data_amdx}()},
\code{\link{getPofZ}()},
\code{\link{get_gxp_both_models}()},
\code{\link{label_spacer}()},
\code{\link{plot_loc}()},
\code{\link{reformat_run_name}()}
}
\concept{Figure SX}
