% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SF1.R
\name{theme_hyb}
\alias{theme_hyb}
\title{Custom ggplot theme}
\usage{
theme_hyb(legend.position = "none", ...)
}
\arguments{
\item{legend.position}{string, according to ggplot2::theme()}

\item{...}{arguments funneled through to ggplot2::theme()}
}
\description{
\code{theme_hyb} ggplot theme for hybridization plots
}
\seealso{
Other Suppl Figure 1: 
\code{\link{getPofZ}()},
\code{\link{label_spacer}()},
\code{\link{plot_loc}()}
}
\concept{Suppl Figure 1}
