% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_f1.R
\name{whisker_points}
\alias{whisker_points}
\title{Whisker points}
\usage{
whisker_points(x)
}
\arguments{
\item{x}{input vector}
}
\description{
\code{whisker_points} calculates the extent of boxplot whiskers.
}
\details{
This is used to plot boxplots without outliers.
}
\seealso{
Other Figure 1: 
\code{\link{make_faint_clr}()},
\code{\link{network_layout}()},
\code{\link{plot_fish}()},
\code{\link{plot_network}()},
\code{\link{summarize_dxy}()},
\code{\link{summarize_fst}()}
}
\concept{Figure 1}
